requirements.txt
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.12.4
django-on-heroku==1.1.2
drf-yasg==1.20.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
requests==2.25.1